+++ server/package.json
@@ -3,9 +3,10 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "output_path": "../public/dummy",
   "scripts": {
     "build": "tsc ServerMain.ts",
-    "start": "ts-node ServerMain.ts"
+    "start": "EXPORT_DIR=\"../src/utils\" ts-node ServerMain.ts"
   },
   "author": "",
   "license": "ISC",
