+++ src/circuit/BothScript/CircuitInfo.ts
@@ -43,40 +43,6 @@ export class CircuitInfoUtils {
     })
     return [basisInfo, partsInfos, wiresInfos]
   }
-
-  static infoToJson(
-    circuitBasisInfo: CircuitBasisInfo,
-    circuitPartsInfos: CircuitPartsInfo[],
-    circuitWiresInfos: CircuitWiresInfo[]
-  ): string {
-    const objs = new Array<ICircuit>(0)
-    const basisObj: IBasisInfo = {
-      class: 'basis',
-      sizeX: circuitBasisInfo.sizeX,
-      sizeY: circuitBasisInfo.sizeY
-    }
-    objs.push(basisObj)
-    circuitPartsInfos.forEach(v => {
-      const obj: IPartsInfo = {
-        class: 'parts',
-        x: v.x,
-        z: v.z,
-        isBig: v.isBig,
-        category: v.category
-      }
-      objs.push(obj)
-    })
-    circuitWiresInfos.forEach(v => {
-      const obj: IWiresInfo = {
-        class: 'wires',
-        x: v.x,
-        z: v.z,
-        wires: v.wires
-      }
-      objs.push(obj)
-    })
-    return JSON.stringify(objs)
-  }
 }
 
 interface ICircuit {
