+++ src/utils/circuit.json
@@ -1,0 +1,1 @@
+[{"class":"basis","sizeX":50,"sizeY":50},{"class":"parts","x":24,"z":24,"isBig":false,"category":"None"},{"class":"parts","x":23,"z":23,"isBig":false,"category":"Web"},{"class":"parts","x":26,"z":23,"isBig":false,"category":"Reversing"},{"class":"wires","x":27,"z":21,"wires":[[0,-1],[1,-1],[2,-1],[3,-1],[4,-1]]},{"class":"wires","x":21,"z":22,"wires":[[11,-1]]},{"class":"wires","x":22,"z":22,"wires":[[11,28]]},{"class":"wires","x":23,"z":22,"wires":[[3,28],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1]]},{"class":"wires","x":24,"z":22,"wires":[[3,28],[4,27],[5,26],[6,25],[7,24],[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1]]},{"class":"wires","x":25,"z":22,"wires":[[0,-1],[1,-1],[2,-1],[3,-1],[4,-1],[5,-1],[6,-1],[7,-1],[8,31],[9,30],[10,29],[11,28],[12,27],[13,26],[14,25],[15,24]]},{"class":"wires","x":26,"z":22,"wires":[[0,31],[1,30],[2,29],[3,28],[4,27],[5,26],[6,25],[7,24]]},{"class":"wires","x":27,"z":22,"wires":[[0,23],[1,22],[2,21],[3,20],[4,19]]},{"class":"wires","x":28,"z":22,"wires":[[5,-1]]},{"class":"wires","x":24,"z":23,"wires":[[0,31],[1,30],[2,29],[3,20],[4,19],[5,18],[6,17],[7,16],[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1]]},{"class":"wires","x":25,"z":23,"wires":[[0,31],[1,30],[2,29],[3,28],[4,27],[5,26],[6,25],[7,24],[8,23],[9,22],[10,21],[11,20],[12,19],[13,18],[14,17],[15,16]]},{"class":"wires","x":27,"z":23,"wires":[[2,-1],[3,-1],[4,-1],[5,-1],[6,-1],[24,23],[25,22],[26,21],[27,20],[28,19],[29,10],[30,1],[31,0]]},{"class":"wires","x":28,"z":23,"wires":[[29,18]]},{"class":"wires","x":21,"z":24,"wires":[[12,-1]]},{"class":"wires","x":22,"z":24,"wires":[[12,27]]},{"class":"wires","x":23,"z":24,"wires":[[8,7],[9,6],[10,5],[11,4],[12,3],[13,2],[14,1],[15,0],[20,27]]},{"class":"wires","x":27,"z":24,"wires":[[2,21],[3,20],[4,19],[5,18],[6,17],[22,1],[23,0],[24,-1],[25,-1],[26,-1],[27,-1],[28,-1],[29,-1],[30,-1],[31,-1]]},{"class":"wires","x":28,"z":24,"wires":[[7,-1]]},{"class":"wires","x":23,"z":25,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1],[16,15],[17,14],[18,13],[19,12],[20,11],[21,10],[22,9],[23,8]]},{"class":"wires","x":27,"z":25,"wires":[[22,-1],[23,-1],[24,-1],[25,-1],[26,21],[27,20],[28,19],[29,18],[30,17],[31,8]]},{"class":"wires","x":28,"z":25,"wires":[[0,-1],[1,-1],[2,-1],[3,-1],[4,-1],[5,-1],[6,-1],[7,-1],[31,16]]},{"class":"wires","x":24,"z":26,"wires":[[16,7],[17,6],[18,5],[19,4],[20,3],[21,2],[22,1],[23,0]]},{"class":"wires","x":25,"z":26,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1],[16,7],[17,6],[18,5],[19,4],[20,3],[21,2],[22,1],[23,0]]},{"class":"wires","x":27,"z":26,"wires":[[0,-1],[1,-1],[2,-1],[3,-1],[4,-1],[5,-1],[6,-1],[24,15],[25,14],[26,13],[27,12],[28,11],[29,10],[30,9],[31,8]]},{"class":"wires","x":28,"z":26,"wires":[[24,23],[25,22],[26,21],[27,20],[28,19],[29,18],[30,17],[31,16]]},{"class":"wires","x":23,"z":27,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1]]},{"class":"wires","x":24,"z":27,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1],[16,31],[17,30],[18,29],[19,28],[20,27],[21,26],[22,25],[23,24]]},{"class":"wires","x":25,"z":27,"wires":[[8,7],[9,6],[10,5],[11,4],[12,3],[13,2],[14,1],[15,0],[16,31],[17,30],[18,29],[19,28],[20,27],[21,26],[22,25],[23,24]]},{"class":"wires","x":27,"z":27,"wires":[[0,-1],[1,-1],[2,-1],[3,-1],[4,-1],[5,-1],[6,-1],[24,23],[25,22],[26,21],[27,20],[28,19],[29,18],[30,17],[31,8]]},{"class":"wires","x":28,"z":27,"wires":[[31,0]]},{"class":"wires","x":25,"z":28,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1],[14,-1],[15,-1],[16,15],[17,14],[18,13],[19,12],[20,11],[21,10],[22,9],[23,8]]},{"class":"wires","x":27,"z":28,"wires":[[24,23],[25,22],[26,21],[27,20],[28,19],[29,18],[30,17],[31,0]]},{"class":"wires","x":28,"z":28,"wires":[[23,-1]]},{"class":"wires","x":24,"z":29,"wires":[[14,-1],[15,-1]]},{"class":"wires","x":25,"z":29,"wires":[[14,25],[15,24]]},{"class":"wires","x":26,"z":29,"wires":[[16,7],[17,6],[18,5],[19,4],[20,3],[21,2],[22,25],[23,24]]},{"class":"wires","x":27,"z":29,"wires":[[23,0]]},{"class":"wires","x":25,"z":30,"wires":[[8,-1],[9,-1],[10,-1],[11,-1],[12,-1],[13,-1]]},{"class":"wires","x":26,"z":30,"wires":[[16,31],[17,30],[18,29],[19,28],[20,27],[21,26]]},{"class":"wires","x":27,"z":30,"wires":[[23,-1]]}]
